masterpiece
Finally, vignettes done right ! Super easy and enjoyable  to use.
On top of that, improvements on readers and writers structure and inheritance. As well as a new polymorphic TransformingReader class yielder that can perform operation on frames at read time and be fed to vignettes, converters, etc as a normal reader, seamlessly.